Remove obsolete jenkins playbook (#315)
The jenkins playbook and its associated file have been removed as they are no
longer needed. This playbook was responsible for applying the jenkins role to
the specified hosts. However, the role itself is no longer used in the project,
so the playbook has been deleted to clean up the codebase.

Signed-off-by: Christian Berendt <berendt@osism.tech>