Fix monitoring issue described in ansible-playbooks/#392 (#401)
[PR#392](https://github.com/osism/ansible-playbooks/pull/392) added
Kubernetes monitoring deployment play. Kubernetes monitoring is
intended to monitor the k3s instance control plane components out of
the box.
PR#392 observes several issues that are mentioned in the
PR's descriptioni, which are related to the k3s distribution.

This commit  addresses the 1. issue where the monitoring solution
could not monitor K3s control plane components kube-scheduler,
kube-proxy, and kube-controller-manager.

The second part of the fix: osism/defaults#177

Signed-off-by: Matej Feder <matej.feder@dnation.cloud>