Always fetch keys in ceph-copy-keys when explicitly tagged
Currently using the `fetch` tag for `ceph-copy-keys` results in the task
being skipped since the fact used in the `when` condition is unset and
defaults to `false`.
Therefore the condition is changed, so that the task is also run when it
is explicitly tagged.

Part of https://github.com/osism/issues/issues/1218

Signed-off-by: Jan Horstmann <horstmann@osism.tech>